+++ packages/ui/src/index.ts
@@ -17,6 +17,7 @@ export {
 	type WikiTarget,
 } from "./editor/EditorView";
 export { FormattingStatusBar } from "./editor/FormattingStatusBar";
+export { classifyHref } from "./editor/href";
 export { LinkCreationGhostExtension } from "./editor/LinkCreationGhostExtension";
 export { SmartLinkExtension } from "./editor/SmartLinkExtension";
 export { VirtualCursor } from "./editor/VirtualCursor";